Development Map
Developement for all provinces done.

Uncolonized all land north and east of the lunar walls on the moon for the Lunar Capital.

Colonized all land south of the lunar walls on the moon for the Lunar Capital.

Swapped a few lunar urban terrain type on the moon.

Fixed the lunar mountain gfx not working.

Added more variety of trade goods on the moon.